break out synced client and concurrent change sync service tests
