+++ core/tests/sync_service_tests.rs
@@ -4,132 +4,6 @@ use lockbook_core::Core;
 use test_utils::*;
 
 /*  ---------------------------------------------------------------------------------------------------------------
-    Tests that operate on one device and sync
-    (work should be none, deleted files should be pruned)
----------------------------------------------------------------------------------------------------------------  */
-
-#[test]
-fn synced_device() {
-    let assert_stuff = |core: &Core| {
-        core.validate().unwrap();
-        assert_local_work_paths(&core, &[]);
-        assert_server_work_paths(&core, &[]);
-        assert_deleted_files_pruned(&core);
-        assert_new_synced_client_dbs_eq(&core);
-    };
-
-    // unmodified
-    {
-        let core = test_core_with_account();
-        core.sync(None).unwrap();
-        assert_all_paths(&core, &["/"]);
-        assert_all_document_contents(&core, &[]);
-        assert_stuff(&core);
-    }
-
-    // new_file
-    {
-        let core = test_core_with_account();
-        core.create_at_path("/document").unwrap();
-        core.sync(None).unwrap();
-        assert_all_paths(&core, &["/", "/document"]);
-        assert_all_document_contents(&core, &[("/document", b"")]);
-        assert_stuff(&core);
-    }
-
-    // new_file_name_same_as_username
-    {
-        let core = test_core_with_account();
-        let account = core.get_account().unwrap();
-        core.create_at_path(&format!("/{}", &account.username))
-            .unwrap();
-        core.sync(None).unwrap();
-        let account = core.get_account().unwrap();
-        let document_path = format!("/{}", account.username);
-        assert_all_paths(&core, &["/", &document_path]);
-        assert_all_document_contents(&core, &[(&document_path, b"")]);
-        assert_stuff(&core);
-    }
-
-    // new_files
-    {
-        let core = test_core_with_account();
-        core.create_at_path("/a/b/c/d").unwrap();
-        core.sync(None).unwrap();
-        assert_all_paths(&core, &["/", "/a/", "/a/b/", "/a/b/c/", "/a/b/c/d"]);
-        assert_all_document_contents(&core, &[("/a/b/c/d", b"")]);
-        assert_stuff(&core);
-    }
-
-    // edited_document
-    {
-        let core = test_core_with_account();
-        core.create_at_path("/document").unwrap();
-        write_path(&core, "/document", b"document content").unwrap();
-        core.sync(None).unwrap();
-        assert_all_paths(&core, &["/", "/document"]);
-        assert_all_document_contents(&core, &[("/document", b"document content")]);
-        assert_stuff(&core);
-    }
-
-    // move
-    {
-        let core = test_core_with_account();
-        let folder = core.create_at_path("/folder/").unwrap();
-        let doc = core.create_at_path("/document").unwrap();
-        core.move_file(doc.id, folder.id).unwrap();
-        core.sync(None).unwrap();
-        assert_all_paths(&core, &["/", "/folder/", "/folder/document"]);
-        assert_all_document_contents(&core, &[("/folder/document", b"")]);
-        assert_stuff(&core);
-    }
-
-    // rename
-    {
-        let core = test_core_with_account();
-        let doc = core.create_at_path("/document").unwrap();
-        core.rename_file(doc.id, "document2").unwrap();
-        core.sync(None).unwrap();
-        assert_all_paths(&core, &["/", "/document2"]);
-        assert_all_document_contents(&core, &[("/document2", b"")]);
-        assert_stuff(&core);
-    }
-
-    // delete
-    {
-        let core = test_core_with_account();
-        let doc = core.create_at_path("/document").unwrap();
-        core.delete_file(doc.id).unwrap();
-        core.sync(None).unwrap();
-        assert_all_paths(&core, &["/"]);
-        assert_all_document_contents(&core, &[]);
-        assert_stuff(&core);
-    }
-
-    // delete_parent
-    {
-        let core = test_core_with_account();
-        core.create_at_path("/folder/document").unwrap();
-        delete_path(&core, "/folder/").unwrap();
-        core.sync(None).unwrap();
-        assert_all_paths(&core, &["/"]);
-        assert_all_document_contents(&core, &[]);
-        assert_stuff(&core);
-    }
-
-    // delete_grandparent
-    {
-        let core = test_core_with_account();
-        core.create_at_path("/grandparent/parent/document").unwrap();
-        delete_path(&core, "/grandparent/").unwrap();
-        core.sync(None).unwrap();
-        assert_all_paths(&core, &["/"]);
-        assert_all_document_contents(&core, &[]);
-        assert_stuff(&core);
-    }
-}
-
-/*  ---------------------------------------------------------------------------------------------------------------
     Tests that operate on one device after syncing
 ---------------------------------------------------------------------------------------------------------------  */
 
@@ -876,873 +750,6 @@ fn synced_change_new_synced_device() {
         assert_all_paths(&c2, &["/"]);
         assert_all_document_contents(&c2, &[]);
         assert_stuff(&c1, &c2);
-    }
-}
-
-/*  ---------------------------------------------------------------------------------------------------------------
-    Tests that setup two synced devices, operate on both devices, then sync both twice
-    (work should be none, devices dbs should be equal, deleted files should be pruned)
----------------------------------------------------------------------------------------------------------------  */
-
-#[test]
-fn concurrent_change() {
-    let sync_and_assert = |c1: &Core, c2: &Core| {
-        c1.sync(None).unwrap();
-        c2.sync(None).unwrap();
-        c1.sync(None).unwrap();
-        c2.sync(None).unwrap();
-
-        c1.validate().unwrap();
-        assert_dbs_eq(c1, c2);
-        assert_local_work_paths(c1, &[]);
-        assert_server_work_paths(c1, &[]);
-        assert_deleted_files_pruned(c1);
-    };
-
-    // identical_move
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.create_at_path("/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        move_by_path(&c1, "/document", "/parent/").unwrap();
-        move_by_path(&c2, "/document", "/parent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/parent/", "/parent/document"]);
-        assert_all_document_contents(&c2, &[("/parent/document", b"")]);
-    }
-
-    // different_move
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.create_at_path("/parent2/").unwrap();
-        c1.create_at_path("/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        move_by_path(&c1, "/document", "/parent/").unwrap();
-        move_by_path(&c2, "/document", "/parent2/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/parent/", "/parent2/", "/parent/document"]);
-        assert_all_document_contents(&c2, &[("/parent/document", b"")]);
-    }
-
-    // identical_rename
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        rename_path(&c1, "/document", "/document2").unwrap();
-        rename_path(&c2, "/document", "/document2").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/document2"]);
-        assert_all_document_contents(&c2, &[("/document2", b"")]);
-    }
-
-    // different_rename
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        rename_path(&c1, "/document", "/document2").unwrap();
-        rename_path(&c2, "/document", "/document3").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/document2"]);
-        assert_all_document_contents(&c2, &[("/document2", b"")]);
-    }
-
-    // move_then_rename
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.create_at_path("/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        move_by_path(&c1, "/document", "/parent/").unwrap();
-        rename_path(&c2, "/document", "/document2").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/parent/", "/parent/document2"]);
-        assert_all_document_contents(&c2, &[("/parent/document2", b"")]);
-    }
-
-    // rename_then_move
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.create_at_path("/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        rename_path(&c1, "/document", "/document2").unwrap();
-        move_by_path(&c2, "/document", "/parent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/parent/", "/parent/document2"]);
-        assert_all_document_contents(&c2, &[("/parent/document2", b"")]);
-    }
-
-    // identical_delete
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/document").unwrap();
-        delete_path(&c2, "/document").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // identical_delete_parent
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/parent/").unwrap();
-        delete_path(&c2, "/parent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // delete_parent_then_direct
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/parent/").unwrap();
-        delete_path(&c2, "/parent/document").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // delete_direct_then_parent
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/parent/document").unwrap();
-        delete_path(&c2, "/parent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // identical_delete_grandparent
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/grandparent/parent/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/grandparent/").unwrap();
-        delete_path(&c2, "/grandparent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // delete_grandparent_then_direct
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/grandparent/parent/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/grandparent/").unwrap();
-        delete_path(&c2, "/grandparent/parent/document").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // delete_direct_then_grandparent
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/grandparent/parent/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/grandparent/parent/document").unwrap();
-        delete_path(&c2, "/grandparent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // delete_grandparent_then_parent
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/grandparent/parent/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/grandparent/").unwrap();
-        delete_path(&c2, "/grandparent/parent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // delete_parent_then_grandparent
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/grandparent/parent/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/grandparent/parent/").unwrap();
-        delete_path(&c2, "/grandparent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // move_then_delete
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.create_at_path("/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        move_by_path(&c1, "/document", "/parent/").unwrap();
-        delete_path(&c2, "/document").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/parent/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // delete_then_move
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.create_at_path("/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/document").unwrap();
-        move_by_path(&c2, "/document", "/parent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/parent/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // move_then_delete_new_parent
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.create_at_path("/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        move_by_path(&c1, "/document", "/parent/").unwrap();
-        delete_path(&c2, "/parent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // delete_new_parent_then_move
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.create_at_path("/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/parent/").unwrap();
-        move_by_path(&c2, "/document", "/parent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // move_then_delete_old_parent
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        move_by_path(&c1, "/parent/document", "").unwrap();
-        delete_path(&c2, "/parent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/document"]);
-        assert_all_document_contents(&c2, &[("/document", b"")]);
-    }
-
-    // delete_old_parent_then_move
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/parent/").unwrap();
-        move_by_path(&c2, "/parent/document", "").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // rename_then_delete
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        rename_path(&c1, "/document", "/document2").unwrap();
-        delete_path(&c2, "/document").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // delete_then_rename
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/document").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/document").unwrap();
-        rename_path(&c2, "/document", "/document2").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // create_then_move_parent
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.create_at_path("/parent2/").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        c1.create_at_path("/parent/document").unwrap();
-        move_by_path(&c2, "/parent/", "/parent2/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/parent2/", "/parent2/parent/", "/parent2/parent/document"]);
-        assert_all_document_contents(&c2, &[("/parent2/parent/document", b"")]);
-    }
-
-    // move_parent_then_create
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.create_at_path("/parent2/").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        move_by_path(&c1, "/parent/", "/parent2/").unwrap();
-        c2.create_at_path("/parent/document").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/parent2/", "/parent2/parent/", "/parent2/parent/document"]);
-        assert_all_document_contents(&c2, &[("/parent2/parent/document", b"")]);
-    }
-
-    // create_then_rename_parent
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        c1.create_at_path("/parent/document").unwrap();
-        rename_path(&c2, "/parent/", "/parent2/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/parent2/", "/parent2/document"]);
-        assert_all_document_contents(&c2, &[("/parent2/document", b"")]);
-    }
-
-    // rename_parent_then_create
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        rename_path(&c1, "/parent/", "/parent2/").unwrap();
-        c2.create_at_path("/parent/document").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/parent2/", "/parent2/document"]);
-        assert_all_document_contents(&c2, &[("/parent2/document", b"")]);
-    }
-
-    // create_then_delete_parent
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        c1.create_at_path("/parent/document").unwrap();
-        delete_path(&c2, "/parent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // delete_parent_then_create
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/parent/").unwrap();
-        c2.create_at_path("/parent/document").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // create_then_delete_grandparent
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/grandparent/parent/").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        c1.create_at_path("/grandparent/parent/document").unwrap();
-        delete_path(&c2, "/grandparent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // delete_grandparent_then_create
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/grandparent/parent/").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/grandparent/").unwrap();
-        c2.create_at_path("/grandparent/parent/document").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // identical_content_edit_not_mergable
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/document.draw").unwrap();
-        write_path(&c1, "/document.draw", b"document content").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        write_path(&c1, "/document.draw", b"document content 2").unwrap();
-        write_path(&c2, "/document.draw", b"document content 2").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/document.draw", "/document-1.draw"]);
-        assert_all_document_contents(
-            &c2,
-            &[
-                ("/document.draw", b"document content 2"),
-                ("/document-1.draw", b"document content 2"),
-            ],
-        );
-    }
-
-    // identical_content_edit_mergable
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/document.md").unwrap();
-        write_path(&c1, "/document.md", b"document content").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        write_path(&c1, "/document.md", b"document content 2").unwrap();
-        write_path(&c2, "/document.md", b"document content 2").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/document.md"]);
-        assert_all_document_contents(&c2, &[("/document.md", b"document content 2")]);
-    }
-
-    // different_content_edit_not_mergable
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/document.draw").unwrap();
-        write_path(&c1, "/document.draw", b"document\n\ncontent\n").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        write_path(&c1, "/document.draw", b"document 2\n\ncontent\n").unwrap();
-        write_path(&c2, "/document.draw", b"document\n\ncontent 2\n").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/document.draw", "/document-1.draw"]);
-        assert_all_document_contents(
-            &c2,
-            &[
-                ("/document.draw", b"document 2\n\ncontent\n"),
-                ("/document-1.draw", b"document\n\ncontent 2\n"),
-            ],
-        );
-    }
-
-    // different_content_edit_mergable
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/document.md").unwrap();
-        write_path(&c1, "/document.md", b"document\n\ncontent\n").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        write_path(&c1, "/document.md", b"document 2\n\ncontent\n").unwrap();
-        write_path(&c2, "/document.md", b"document\n\ncontent 2\n").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/document.md"]);
-        assert_all_document_contents(&c2, &[("/document.md", b"document 2\n\ncontent 2\n")]);
-    }
-
-    // different_content_edit_mergable_with_move_in_first_sync
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.create_at_path("/document.md").unwrap();
-        write_path(&c1, "/document.md", b"document\n\ncontent\n").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        write_path(&c1, "/document.md", b"document 2\n\ncontent\n").unwrap();
-        move_by_path(&c1, "/document.md", "/parent/").unwrap();
-        write_path(&c2, "/document.md", b"document\n\ncontent 2\n").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/parent/", "/parent/document.md"]);
-        assert_all_document_contents(&c2, &[("/parent/document.md", b"document 2\n\ncontent 2\n")]);
-    }
-
-    // different_content_edit_mergable_with_move_in_second_sync
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.create_at_path("/document.md").unwrap();
-        write_path(&c1, "/document.md", b"document\n\ncontent\n").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        write_path(&c1, "/document.md", b"document 2\n\ncontent\n").unwrap();
-        write_path(&c2, "/document.md", b"document\n\ncontent 2\n").unwrap();
-        move_by_path(&c2, "/document.md", "/parent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/parent/", "/parent/document.md"]);
-        assert_all_document_contents(&c2, &[("/parent/document.md", b"document 2\n\ncontent 2\n")]);
-    }
-
-    // move_then_edit_content
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.create_at_path("/document.md").unwrap();
-        write_path(&c1, "/document.md", b"document content").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        move_by_path(&c1, "/document.md", "/parent/").unwrap();
-        write_path(&c2, "/document.md", b"document content 2").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/parent/", "/parent/document.md"]);
-        assert_all_document_contents(&c2, &[("/parent/document.md", b"document content 2")]);
-    }
-
-    // edit_content_then_move
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/").unwrap();
-        c1.create_at_path("/document.md").unwrap();
-        write_path(&c1, "/document.md", b"document content").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        write_path(&c1, "/document.md", b"document content 2").unwrap();
-        move_by_path(&c2, "/document.md", "/parent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/parent/", "/parent/document.md"]);
-        assert_all_document_contents(&c2, &[("/parent/document.md", b"document content 2")]);
-    }
-
-    // rename_then_edit_content
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/document.md").unwrap();
-        write_path(&c1, "/document.md", b"document content").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        rename_path(&c1, "/document.md", "document2.md").unwrap();
-        write_path(&c2, "/document.md", b"document content 2").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/document2.md"]);
-        assert_all_document_contents(&c2, &[("/document2.md", b"document content 2")]);
-    }
-
-    // edit_content_then_rename
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/document.md").unwrap();
-        write_path(&c1, "/document.md", b"document content").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        write_path(&c1, "/document.md", b"document content 2").unwrap();
-        rename_path(&c2, "/document.md", "document2.md").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/", "/document2.md"]);
-        assert_all_document_contents(&c2, &[("/document2.md", b"document content 2")]);
-    }
-
-    // delete_then_edit_content
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/document.md").unwrap();
-        write_path(&c1, "/document.md", b"document content").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/document.md").unwrap();
-        write_path(&c2, "/document.md", b"document content 2").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // edit_content_then_delete
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/document.md").unwrap();
-        write_path(&c1, "/document.md", b"document content").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        write_path(&c1, "/document.md", b"document content 2").unwrap();
-        delete_path(&c2, "/document.md").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // delete_parent_then_edit_content
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/document.md").unwrap();
-        write_path(&c1, "/parent/document.md", b"document content").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/parent/").unwrap();
-        write_path(&c2, "/parent/document.md", b"document content 2").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // edit_content_then_delete_parent
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/parent/document.md").unwrap();
-        write_path(&c1, "/parent/document.md", b"document content").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        write_path(&c1, "/parent/document.md", b"document content 2").unwrap();
-        delete_path(&c2, "/parent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // delete_grandparent_then_edit_content
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/grandparent/parent/document.md")
-            .unwrap();
-        write_path(&c1, "/grandparent/parent/document.md", b"document content").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        delete_path(&c1, "/grandparent/").unwrap();
-        write_path(&c2, "/grandparent/parent/document.md", b"document content 2").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
-    }
-
-    // edit_content_then_delete_grandparent
-    {
-        let c1 = test_core_with_account();
-        c1.create_at_path("/grandparent/parent/document.md")
-            .unwrap();
-        write_path(&c1, "/grandparent/parent/document.md", b"document content").unwrap();
-        c1.sync(None).unwrap();
-
-        let c2 = another_client(&c1);
-        c2.sync(None).unwrap();
-
-        write_path(&c1, "/grandparent/parent/document.md", b"document content 2").unwrap();
-        delete_path(&c2, "/grandparent/").unwrap();
-
-        sync_and_assert(&c1, &c2);
-        assert_all_paths(&c2, &["/"]);
-        assert_all_document_contents(&c2, &[]);
     }
 }
